# docs/api-contract.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Integration Operations Hub Gateway",
    "version": "1.0.0"
  },
  "servers": [{"url": "http://localhost:3000"}],
  "paths": {
    "/api/health": {
      "get": {
        "summary": "Check gateway and upstream health",
        "responses": {"200": {"description": "All dependencies healthy"}, "503": {"description": "One or more dependencies unavailable"}}
      }
    },
    "/api/jobs": {
      "get": {
        "summary": "List integration jobs",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "query", "in": "query", "schema": {"type": "string"}},
          {"name": "limit", "in": "query", "schema": {"type": "integer", "minimum": 1, "maximum": 100}}
        ],
        "responses": {"200": {"description": "A filtered page of jobs"}}
      },
      "post": {
        "summary": "Validate, transform, and synchronize employee records",
        "security": [{"bearerAuth": []}],
        "parameters": [{"name": "X-Correlation-ID", "in": "header", "schema": {"type": "string", "format": "uuid"}}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/CreateJobRequest"}
            }
          }
        },
        "responses": {
          "202": {"description": "Synchronization completed or partially completed", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/IntegrationJob"}}}},
          "400": {"description": "Validation error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
          "502": {"description": "Upstream failure", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
          "503": {"description": "Open upstream circuit", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}}
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {"type": "http", "scheme": "bearer"}
    },
    "schemas": {
      "EmployeeRecord": {
        "type": "object",
        "required": ["employeeNumber", "firstName", "lastName", "email", "department", "status"],
        "properties": {
          "employeeNumber": {"type": "string", "minLength": 2, "maxLength": 20},
          "firstName": {"type": "string", "maxLength": 80},
          "lastName": {"type": "string", "maxLength": 80},
          "email": {"type": "string", "format": "email"},
          "department": {"type": "string", "maxLength": 80},
          "status": {"type": "string", "enum": ["ACTIVE", "INACTIVE"]}
        }
      },
      "CreateJobRequest": {
        "type": "object",
        "required": ["sourceSystem", "targetSystem", "records"],
        "properties": {
          "sourceSystem": {"type": "string"},
          "targetSystem": {"type": "string"},
          "records": {"type": "array", "minItems": 1, "maxItems": 100, "items": {"$ref": "#/components/schemas/EmployeeRecord"}}
        }
      },
      "IntegrationJob": {
        "type": "object",
        "required": ["id", "correlationId", "sourceSystem", "targetSystem", "status", "recordsReceived", "recordsSucceeded", "recordsFailed", "createdAt"],
        "properties": {
          "id": {"type": "string", "format": "uuid"},
          "correlationId": {"type": "string", "format": "uuid"},
          "sourceSystem": {"type": "string"},
          "targetSystem": {"type": "string"},
          "status": {"type": "string", "enum": ["QUEUED", "RUNNING", "SUCCEEDED", "PARTIAL", "FAILED"]},
          "recordsReceived": {"type": "integer"},
          "recordsSucceeded": {"type": "integer"},
          "recordsFailed": {"type": "integer"},
          "createdAt": {"type": "string", "format": "date-time"},
          "completedAt": {"type": "string", "format": "date-time"},
          "errorMessage": {"type": "string"}
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "required": ["code", "message", "correlationId"],
        "properties": {
          "code": {"type": "string"},
          "message": {"type": "string"},
          "correlationId": {"type": "string"},
          "details": {}
        }
      }
    }
  }
}